firmware: move clock generator logic to clock_gen.h|c
