+++ firmware/common/hackrf_core.h
@@ -31,7 +31,6 @@ extern "C" {
 #include <stdint.h>
 
 #include "cpld_jtag.h"
-#include "fixed_point.h"
 #include "i2c_bus.h"
 #include "i2c_lpc.h"
 #include "max283x.h"
@@ -69,8 +68,6 @@ extern rf_path_t rf_path;
 extern jtag_t jtag_cpld;
 extern i2c_bus_t i2c0;
 
-void clock_gen_init(void);
-void clock_gen_shutdown(void);
 void ssp1_set_mode_max283x(void);
 void ssp1_set_mode_max5864(void);
 #ifdef IS_PRALINE
@@ -88,10 +85,6 @@ void disable_3v3aux_power(void);
 #endif
 void enable_1v8_power(void);
 void disable_1v8_power(void);
-
-fp_28_36_t sample_rate_set(const fp_28_36_t sample_rate, const bool program);
-
-clock_source_t activate_best_clock_source(void);
 
 #if defined(IS_RAD1O) || defined(IS_HACKRF_ONE) || defined(IS_PRALINE)
 void enable_rf_power(void);
@@ -116,32 +109,12 @@ void halt_and_flash(const uint32_t duration);
 
 #ifdef IS_PRALINE
 typedef enum {
-	P1_SIGNAL_TRIGGER_IN = 0,
-	P1_SIGNAL_AUX_CLK1 = 1,
-	P1_SIGNAL_CLKIN = 2,
-	P1_SIGNAL_TRIGGER_OUT = 3,
-	P1_SIGNAL_P22_CLKIN = 4,
-	P1_SIGNAL_P2_5 = 5,
-	P1_SIGNAL_NC = 6,
-	P1_SIGNAL_AUX_CLK2 = 7,
-} p1_ctrl_signal_t;
-
-typedef enum {
-	P2_SIGNAL_CLK3 = 0,
-	P2_SIGNAL_TRIGGER_IN = 2,
-	P2_SIGNAL_TRIGGER_OUT = 3,
-} p2_ctrl_signal_t;
-
-typedef enum {
 	CLKIN_SIGNAL_P1 = 0,
 	CLKIN_SIGNAL_P22 = 1,
 } clkin_signal_t;
 
-void p1_ctrl_set(const p1_ctrl_signal_t signal);
-void p2_ctrl_set(const p2_ctrl_signal_t signal);
 void narrowband_filter_set(const uint8_t value);
 void clkin_ctrl_set(const clkin_signal_t value);
-void pps_out_set(const uint8_t value);
 #endif
 
 #ifdef __cplusplus
