+++ firmware/common/ui_portapack.c
@@ -26,8 +26,8 @@
 #include <stddef.h>
 #include <stdint.h>
 
+#include "clock_gen.h"
 #include "portapack.h"
-#include "radio.h"
 #include "rf_path.h"
 #include "transceiver_mode.h"
