+++ firmware/hackrf_usb/hackrf_usb.c
@@ -29,6 +29,7 @@
 #include <libopencm3/lpc43xx/ipc.h>
 
 #include <clkin.h>
+#include <clock_gen.h>
 #include <cpu_clock.h>
 #include <da7219.h>
 #include <delay.h>
