+++ firmware/hackrf_usb/usb_api_board_info.c
@@ -26,6 +26,7 @@
 
 #include <libopencm3/lpc43xx/wwdt.h>
 
+#include <clock_gen.h>
 #include <firmware_info.h>
 #include <hackrf_core.h>
 #include <platform_detect.h>
