+++ firmware/hackrf_usb/usb_api_praline.c
@@ -23,6 +23,7 @@
 
 #include "usb_api_praline.h"
 
+#include <clock_gen.h>
 #include <hackrf_core.h>
 #include <platform_detect.h>
 #include <usb_queue.h>
